pyproject.toml
[project]
name = "weather-recognition"
version = "0.1.0"
description = "Image classification challenge: Weather-Image-Recognition"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "fire>=0.7.1",
    "pre-commit>=4.5.0",
]

[tool.ruff]
line-length = 100
format = { docstring-code-format = true }

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-builtins
    "A",
    # flake8-return
    "RET",
    # flake8-use-pathlib
    "PTH",
    # flynt
    "FLY",
    # mccabe
    "C90",
    # NumPy-specific rules
    "NPY",
    # pep8-naming
    "N",
    # Ruff-specific rules
    "RUF",
    # Pylint - very strict, may be use when have enough time
    "PL",
    # refurb - also may be overwhelming
    "FURB",
]

pycodestyle = { max-doc-length = 105 }

# [tool.ruff.lint.per-file-ignores]
# "__init__.py" = ["E402"]
# "**/{tests,docs,tools}/*" = ["E402"]